+++ src/Resources/Resources/ActiveDirectory/Cmdlets/UpdateAzureADApplicationCommand.cs
@@ -101,7 +101,7 @@ namespace Microsoft.Azure.Commands.ActiveDirectory
                 {
                     DisplayName = DisplayName,
                     Homepage = HomePage,
-                    IdentifierUris = IdentifierUri,
+                    IdentifierUris = (IdentifierUri == null) ? new string[] { } : IdentifierUri,
                     ReplyUrls = ReplyUrl,
                     AvailableToOtherTenants = AvailableToOtherTenants
                 };