+++ src/Resources/Resources/ChangeLog.md
@@ -19,7 +19,11 @@
 -->
 ## Upcoming Release
 
-## Version 1.10.0
+## Version 0.12.0
+* Set `-IdentifierUris` as alias of `-IdentifierUri` in `New-AzADApplication` and changed it as optional parameter
+* Removed default DisplayName of ADServicePrincipal when it is not specified
+
+## Version 0.11.0
 * Release for Azure Stack 2020-09-01-hybrid profile
 
 ## Version 0.10.0-preview